primary refactoring
