+++ mypy.ini
@@ -1,3 +1,4 @@
 [mypy]
 mypy_path = $MYPY_CONFIG_FILE_DIR
 files = $MYPY_CONFIG_FILE_DIR/services/**/*.py
+ignore_missing_imports = True
