+++ services/etl/schemas/elasticsearch.py
